backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.27",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.46.0",
]
